Backend/requirements.txt
fastapi 
uvicorn
sqlalchemy
pypdf
langchain
langchain-community
openai
langchain-openai
python-dotenv
chromadb
